support data beyond 3d for interp tuning support int compression

+++ test/sz_interp_int.cpp
@@ -14,13 +14,13 @@ int main(int argc, char **argv) {
     double reb = atof(argv[argp++]);
     if (argp >= argc) {
         if (dim == 1) {
-            interp_lorenzo_tuning<float, 1>(argv[1], reb, false, dims[0]);
+            interp_lorenzo_tuning<int, 1>(argv[1], reb, false, dims[0]);
         } else if (dim == 2) {
-            interp_lorenzo_tuning<float, 2>(argv[1], reb, false, dims[0], dims[1]);
+            interp_lorenzo_tuning<int, 2>(argv[1], reb, false, dims[0], dims[1]);
         } else if (dim == 3) {
-            interp_lorenzo_tuning<float, 3>(argv[1], reb, true, dims[0], dims[1], dims[2]);
+            interp_lorenzo_tuning<int, 3>(argv[1], reb, true, dims[0], dims[1], dims[2]);
         } else if (dim == 4) {
-            interp_lorenzo_tuning<float, 4>(argv[1], reb, false, dims[0], dims[1], dims[2], dims[3]);
+            interp_lorenzo_tuning<int, 4>(argv[1], reb, false, dims[0], dims[1], dims[2], dims[3]);
         }
         return 0;
     }
@@ -44,21 +44,21 @@ int main(int argc, char **argv) {
     }
     size_t num = 0;
 
-    auto data = SZ::readfile<float>(argv[1], num);
+    auto data = SZ::readfile<int>(argv[1], num);
     double eb = reb * SZ::data_range(data.get(), num);
 
 
     if (dim == 1) {
-        interp_compress_decompress<float, 1>(argv[1], data.get(), num, eb, interp_op, direction_op, block_size,
+        interp_compress_decompress<int, 1>(argv[1], data.get(), num, eb, interp_op, direction_op, block_size,
                                              interp_block_size, dims[0]);
     } else if (dim == 2) {
-        interp_compress_decompress<float, 2>(argv[1], data.get(), num, eb, interp_op, direction_op, block_size,
+        interp_compress_decompress<int, 2>(argv[1], data.get(), num, eb, interp_op, direction_op, block_size,
                                              interp_block_size, dims[0], dims[1]);
     } else if (dim == 3) {
-        interp_compress_decompress<float, 3>(argv[1], data.get(), num, eb, interp_op, direction_op, block_size,
+        interp_compress_decompress<int, 3>(argv[1], data.get(), num, eb, interp_op, direction_op, block_size,
                                              interp_block_size, dims[0], dims[1], dims[2]);
     } else if (dim == 4) {
-        interp_compress_decompress<float, 4>(argv[1], data.release(), num, eb, interp_op,
+        interp_compress_decompress<int, 4>(argv[1], data.release(), num, eb, interp_op,
                                              direction_op, block_size,
                                              interp_block_size, dims[0], dims[1], dims[2], dims[3]);
     }
